+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 [中文](./CHANGELOG_zh-cn.md)
 
+- [v1.4.0](#v140)
 - [v1.3.8](#v138)
 - [v1.3.7](#v137)
 - [v1.3.6](#v136)
@@ -14,6 +15,11 @@
 - [v1.2.1](#v121)
 - [v1.2.0](#v120)
 - [v1.1.1](#v111)
+
+## v1.4.0
+
+- Allow passing parameter with a single slice or array for Get, Set, Append, Insert, Delete methods.
+- Add GreaterThan, LessThan, GreaterThanOrEqual, LessThanOrEqual methods for number comparation.
 
 ## v1.3.8